Split actions into separate files
