+++ frontend/app/lib/actions/index.ts
@@ -1,0 +1,4 @@
+export { updateVocabItem, type State } from './updateVocabularyItem';
+export { createVocabItemGroup } from './createVocabItemGroup';
+export { addVocabItemsFromFiles } from './addVocabItemsFromFiles';
+
